Move WebhookTestCase class from test_hooks.py to test_helpers.py.
